FSR-863 update dependancies (#17)
* https://eaflood.atlassian.net/browse/FSR-863

* Update node version to LTS version

* Add minor range qualifier to package versions

* using version ranges rather than specific versions is considered best
  practice

* Upgrade serverless to v3

Followed upgrade instructions here:
https://www.serverless.com/framework/docs/guides/upgrading-v3

Required a missing env var to be added and also the lambdaHashingVersion
to be specified. After this there were no deprecation warnings.

* Update test data

Note that the key changes frequently and we therefore need a better way
of doing this. Perhaps have a dedicated bucket for test which only
changes on demand.

* Update serverless to use Node 16

* Update xml2js